[TEMP] neutrino: Regenerate and savedefconfig (5.15.154)
Update for prerelease testing: "5.15.154-NeutrinoKernel".

As ethically questionable as this feels, I'm confident Sultan would
want his kernel differentiated from the shenanigans going on over
here, and it is for that reason only I'm slapping my name on it.

Regardless, this kernel is a personal project unlikely to see the
light of day publicly.

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>