[TEMP] neutrino: Regenerate and savedefconfig (5.15.153)
Update for prerelease testing: "5.15.153-NeutrinoKernel".

This feels ethically questionable, but I'm confident Sultan would
want his kernel differentiated from the shenanigans going on over
here, and it is for that reason only that I'm slapping my name on
it.

Not like this kernel is ever seeing the light of day publicly.

[@0ctobot: Enable KSU subsequent to
4532ec1 ("kernelsu: Don't enable KSU by default")]
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>