fs: Manually integrate KernelSU calls
This kernel is in the unique position of technically being a GKI
that has been de-GKI'ed and lacks MODULE support. As a result,
newer methods of automatic integration by way of exploiting KPROBES
and LKM will not work. Thus, we'll need to take the original
approach of hard coding calls to KernelSU as outlined here:

https://kernelsu.org/guide/how-to-integrate-for-non-gki.html

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>